+++ RELEASE_NOTES.md
@@ -8,10 +8,12 @@
    `spec.security.certsSecretPrefix` to specify the secret name containing the certificate for Database. Make sure to create the secret containing the certificates accordingly. 
 * `spec.podSpec.cpu` and `spec.podSpec.memory` has been removed to override the CPU/Memory resources for the 
    database pod, you need to override them using the statefulset spec override under `spec.podSpec.podTemplate.spec.containers`.   
+* Custom labels specified under `metadata.labels` is propagated to the database StatefulSet and the PVC objects.
 
 ## MongoDBOpsManager Resource
 * `spec.applicationDatabase.security.tls.secretRef.name` has been removed. It was deprecated in operator version `v1.10.0`. Please use the field
 `spec.applicationDatabase.security.certsSecretPrefix` to specify the secret name containing the certificate for AppDB. Make sure to create the secret containing the certificates accordingly.
+* * Custom labels specified under `metadata.labels` is propagated to the OM, AppDB and BackupDaemon StatefulSet and the PVC objects.
 
 ## MongoDBUser Resource
 * Changes: